303 subtarefa aprovar orcamento entrega bens consumo instalador entrega garantia instalador

See merge request ar-clube/ar-clube!103

+++ Global/Infra/InfraBanco/Modelos/TorcamentoCotacao.cs
@@ -109,7 +109,7 @@ namespace InfraBanco.Modelos
 
         public DateTime? GarantiaIndicadorDtHrUltAtualiz { get; set; }
 
-        public int? EtgImediataIdTipoUsuarioContexto { get; set; }
+        public short? EtgImediataIdTipoUsuarioContexto { get; set; }
 
         public int? EtgImediataIdUsuarioUltAtualiz { get; set; }
